bug: tried getting the fish animations working
only to realize that every object in my blender file has an animation on it for some reason, which is a bug in my blender work, so I'll come back to this at a later time. I legit learned maybe two hours worth of blender basics and just dove in so I'm sure there will be more mistakes but I'll learn as I go.